chore: Update html and css
- Add styling for form buttons and error message
- Define error message appearance and positioning
- Utilize custom CSS variables for consistent styling

Styles for the .form-btns and .error-message classes were added to enhance the visual presentation and user experience of the form component. Flexbox is used to create a flexible layout for form buttons, and the error message style includes defined font, color, and padding properties for a clear and appealing display. Custom CSS variables were employed for consistent theming.